# payload/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Egregore Projection Plane",
    "version": "0.6.0-phase1"
  },
  "paths": {
...
    "/ingest": {
      "post": {
        "tags": [
          "ANCHORUM Ingest"
        ],
        "summary": "Ingest Event",
        "description": "Accept a Stage-4 IngestionEvent from the ANCHORUM connector.\n\nEvents are staged to the report directory and can be fused into a case\nreport by triggering ``POST /api/v1/anchorum/batch`` with the dossier root.\n\nA lightweight per-source rate limit is applied because this endpoint is\nintentionally public. Production deployments should add source validation\nor a connector token.",
        "operationId": "ingest_event_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestionEvent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestReceipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/login": {
      "get": {
        "summary": "Login Page",
        "description": "Self-hosted login page: enter API key to receive a session cookie.",
        "operationId": "login_page_dashboard_login_get",
        "parameters": [
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Login Submit",
        "description": "Validate API key and set session cookie.",
        "operationId": "login_submit_dashboard_login_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_submit_dashboard_login_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_login_submit_dashboard_login_post": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key"
          }
        },
        "type": "object",
        "required": [
          "api_key"
        ],
        "title": "Body_login_submit_dashboard_login_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestReceipt": {
        "properties": {
          "receipt_id": {
            "type": "string",
            "title": "Receipt Id"
          },
          "epistemic_state": {
            "type": "string",
            "title": "Epistemic State"
          },
          "accepted": {
            "type": "boolean",
            "title": "Accepted"
          },
          "event_id": {
            "type": "string",
            "title": "Event Id"
          }
        },
        "type": "object",
        "required": [
          "receipt_id",
          "epistemic_state",
          "accepted",
          "event_id"
        ],
        "title": "IngestReceipt"
      },
      "IngestionEvent": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "anchorum"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose",
            "default": "legal_case_relevant"
          },
          "source_legality": {
            "type": "string",
            "title": "Source Legality",
            "default": "user_provided"
          },
          "scope_status": {
            "type": "string",
            "title": "Scope Status",
            "default": "case_actor"
          },
          "entity_kind": {
            "type": "string",
            "title": "Entity Kind",
            "default": "email"
          },
          "entity_value": {
            "type": "string",
            "title": "Entity Value",
            "default": ""
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "IngestionEvent",
        "description": "Loose model for the ANCHORUM Stage-4 IngestionEvent.\n\nExtra fields are allowed so the connector payload (which evolves) never\nbreaks the receipt endpoint."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}